requirements.txt
ttkbootstrap
pymysql
pandas
openpyxl
python-dotenv